[UI] EA-673396 Use SelectInProjectViewImpl to select item on Project View pane switch
Now that we have a coroutine-based service for selecting items
in the Project View, it seems logical to use it to select the current
item when switching the current pane.

The service had to be modified a bit to allow for null VirtualFile
(extracted from the PSI element now) and null ActionCallback
(as it isn't needed for this use case).

GitOrigin-RevId: a42bd2941a1e9856973f525e328a77f7b74c06df