[workspace model] RUBY-30762 Avoid usage `VersionedEntityStorage` with caching if we have a lot of modifications
The problem where is that we are constantly getting cached value and modifying the storage. It leads to huge about of
collections freeze and copy. Cache in this case is useless because it invalidates on each diff modification.
The main scenario:
1) At library creation we request cached module entity, `VersionedEntityStorage` check that there was a modification
with builder recalculate the value and freeze the storage
2) At the moment of saving LibraryEntity into the storage we see that collection is freezed and making the copy
And it's going in the loop of libraries creation
Similar behavior for other parts of the freeze

GitOrigin-RevId: 8f967a8d83fe380cc8edd3f290f6f4378c32450d