[tabs] Deprecate TableLayout and CompressibleSingleRowLayout
Use inheritors of MultiRowLayout instead.
Also, remove compressible tabs layout from JBEditorTabs, because it is strange that such a common class with a lot of overrides and usages in places that do not refer to editor, use compressible layout depending on editor settings (that can be unexpected). It would be clearer if inheritors will specify it manually.

GitOrigin-RevId: b72b25868cf7ab4438e8c39b9681643d7ce727b5