[UI] IDEA-317789 Fix Select in actions when the context is Project View
After fixing IDEA-304888, the Project View always uses the file opened
in the editor as the context. This makes sense for the Select in Project View
action, but not for other actions.

Fix by moving creation of the editor-related context from SelectInContextImpl
to SelectInAction. Now two contexts are created, and which one is eventually
passed to the target depends on the kind of the target itself, which is determined
by a new marker interface - SelectInTargetPreferringEditorContext.

Fix for both legacy and new implementations, as 231 still uses the legacy one.

GitOrigin-RevId: c06fc7323cb49f541036f9278d100a71c26ed7c9