[UI] IDEA-312482 Only react to caret movements from correct editor in PositionPanel
When PositionPanel was converted to coroutine-based code, it lost its isOurEditor()
method that was inherited from EditorBasedWidget and was responsible from
filtering out events from other windows (amazingly, every PositionPanel is subscribed
to events from ALL IDE editors in ALL frames).

Fortunately, the new implementation already had a similar functionality
in the selection listener, so the fix is simply extracting it into a method
and reusing it for the caret listener as well.

GitOrigin-RevId: ee63bfa5ab19841a50612379ef219f1ab55ec999