[kotlin] fix nested lambda breakpoint skipping
Previously, KotlinLineBreakpointType used a special heuristic to detect if we should stop on a location. This heuristic picked a lambda on a current source position by an ordinal encoded in the breakpoint, and then it tried to match this lambda with the one the breakpoint was put in. However, it turned out that this heuristic picks wrong lambdas in some cases, and simply using the `getContainingMethod` function works everywhere. So this commit deletes the heuristic in favour of the mentioned function.

^KTIJ-25056 Fixed

GitOrigin-RevId: f69a74a4067b354740131c34dfd68f2ac3d8035f