[build scripts] move intellij.platform.boot and its dependencies to a separate boot.jar (IJPL-128)
PathClassLoader is used as a system classloader, so it's better to ensure that its classpath doesn't include unnecessary classes. It becomes crucial when a scheme with a separate bootstrap classloader is used: if the bootstrap classloader is used as a parent for the classloader of remaining part of the platform, and its classpath includes modules whose dependencies aren't in its classpath, it'll fail with NoClassDefFoundError.

GitOrigin-RevId: 11bc79a6198bd44feb6afa24e4c1218decc47792