[new terminal] IDEA-319857 Do not complete selected item on typing some characters (space, '.', ',' etc.)
DefaultCharFilter is not suitable for terminal commands, because it mostly oriented on code. For example, it completes the selected item on typing some characters. The most annoying is completion after typing space.
Also, I modified the logic for parameters completion: the lookup will be closed if user typed the character, that do not match any of available parameters.

GitOrigin-RevId: d68728c2619234cf1c2b8adefb2a15dec288d831