[project file index] iterate over files from custom file sets in ProjectFileIndex::iterateContent
Before it used the content roots of modules as the starting points for iteration, and, therefore, roots of custom file sets contributed via WorkspaceFileIndexContributor weren't processed. It is needed for RIDER-77296, also it should be faster to take roots from the file sets rather than iterating over modules and their content roots.

GitOrigin-RevId: 58c75cc0ed41f655bffc3fc659b79ed6a53bd98d