[inspection] Smarter check for opt-in in rangeUntil inspections
Don't check for opt-in when `rangeUntil` becomes stable.

In case `rangeUntil` function is not found, inspection acts
as if opt-in is not required. Because the most probable reason
why the feature is enabled but `rangeUntil` not found is some
buggy changes in resolution implementation. We don't want to
block inspection in such case.

^KTIJ-25447 Fixed

GitOrigin-RevId: adb482ecb342830b5be61caa94dff97c59eeabb1