IDEA-318686 Fix storing/restoring size in AbstractPopup
There was a bug with the Dimension Service: when the size
is restored, the insets are added to it, and when it's stored,
the insets are subtracted back.

However, when rounded corners are used, the border is actually
removed after the size is restored, and therefore the insets
are only added, but not subtracted afterward. This led to
the size increasing every time the popup is shown and hidden.

GitOrigin-RevId: b20a403e01be321dc99c1577d1eece576f7cb48a