IDEA-304684 Use NBRA for getPsiFiles / getPsiElement(s) in ProjectViewDropTarget
Pack all PSI stuff needed to perform a drop into a single record, fetch it in one go
with an NBRA, and then use that data when needed.

Still some slow ops remaining, most in the update() call, and also some
isValidTargetStuff (checks whether the directory is excluded or whether
a file is in content).

GitOrigin-RevId: f07bf12b656867d53d1d0e876405c1eb77034042