[git-branches-dashboard] select a visible branch when navigating
The same branch name can be used in several roots in the project. This creates ambiguity when navigating to such branch from branches dashboard, since it does not provide a root. Previously, getBranchRow selected the target branch only using VcsGoToRefComparator which uses comparator provided by the corresponding VcsLogProvider. This commit ensures that the selected branch is also actually visible in the table.

GitOrigin-RevId: 9bcb989b9e8c8ecce00c9e0c41ca10da35626029