[build scripts] write proper location to runtime module descriptor for module which is included in several JARs (IJPL-109)
There are some modules (e.g. intellij.idea.ultimate.build.dependencies) which are included in several JARs in the IDE's distribution, and we need to specify the proper location in the module descriptor to ensure that when another product started from the IDE's distribution it won't get excessive JARs in its classpath.

GitOrigin-RevId: 864f0c84b0e2d66e79b582f03e98fbc148a2bf3e