IDEA-319735 Fix Project View auto-switching pane when current element is not selectable
Caused by a recent slow op fix. There was a check that ensured that the current element
is selectable in the new view and only then attempted to select it. After that check was
removed, it now auto-switches to another view if the element is not present in the new view.

Fixed by adding a flag to select in implementation, that performs the same check,
but now in a slow-op-safe manner, inside the same coroutine that is used for selection.

GitOrigin-RevId: ef80696518dd6d9cb7785974b769d8db04c884ae