IDEA-142105 Ability to open or close directory/folder by single clicking
Just setting toggleClickCount to 1 leads to weird behavior: double-clicking
then expands and immediately collapses back because it's treated as two
subsequent clicks. This is due to a weird clickCount % toggleClickCount == 0
check in BasicTreeUI. To get around that, we double-check that it's not just
divisible, but actually equal to the set value.

GitOrigin-RevId: 9996fc92422d4cfa5f23fb72cbe689f6509acce3