[kotlin] Debugger: IR / Debugger: “Backend Internal error: Exception during IR lowering”
When a breakpoint is inside a local class or a local object and 'this' is evaluated, it's important to add a file containing this class or object for the IR evaluator. Otherwise, only the debugger code fragment will be passed and compiler will not have enough data for IR lowering.

^KTIJ-25491

GitOrigin-RevId: b4c1070bd6315e71e3a0389f54eddd92f0404d1b