IDEA-74443 New Project View sorting API based on enum instead of booleans
Using booleans for every sorting order is becoming ugly: there's isManualOrder,
isSortByType, now we add isSortByTime, and then we have to split this into
ascending/descending, so there'll be a lot of booleans.

Introduce an enum instead, so we can replace most of those booleans with
a single enum value plus one boolean (isFoldersAlwaysOnTop).

GitOrigin-RevId: 4b4c8ca56176c41bd1b6517cb3970907eed5e0a4