Add ability to filter classes out of automatic imports
Currently, some actions like refactoring cause imports to happen
as a side-effect, notably for nested classes. For certain classes,
it is almost never correct to import them. An example here is in
Android Studio where R.string, R.values, R.drawable, etc. are
technically nested classes (despite the lowercase letter).

This change creates a new ClassImportFilter interface and associated
Extension Point, which plugins can use to veto the import of particular
classes. It also creates a ClassInfo nested class so that if it turns
out later that an implementation of ClassImportFilter requires additional
information to make this decision, we do not need to change the API and/or
break existing callers.

This attempts to address https://youtrack.jetbrains.com/issue/KTIJ-25503