IDEA-74443 Restore isManualOrder boolean
Turns out that the manual order can be used in combination with
the usual sorting, when items are sorted manually first and then,
those that are equal (or don't provide manual sorting keys)
are sorted between themselves according to the order set.

So we now only replace isSortByType with the enum of three values,
leaving the manual order alone.

GitOrigin-RevId: 3e21b883db18c483fc2db24382e59d36ea94a055