[inspection] Make Enum.values() inspection tests new behaviour
We want the inspection to test new 1.9 behaviour, i.e. the case when
EnumEntries doesn't have opt-in, but has @WasExperimental.
Because of that `-opt-in` flag was removed from test data.
Current LV in idea is 1.8, so we also have to override API_VERSION to 1.9.

Also, API_VERSION=1.8 set to opt-in tests to not fail when idea
LV becomes 1.9.

^KTIJ-25447

GitOrigin-RevId: 3997eb88957ba554d405a08fff443de04b5f987a