IDEA-320506 Fix details in Project View for Scope views
It didn't work because the code adding inplace comments check
that the parent is a PsiDirectory. For Scope-based views it doesn't
work because it's not a PsiDirectory but a ProjectFileNode instead.

Fix by checking for that one as well and checking that the underlying
file is actually a directory.

Now it apparently only doesn't work for the Packages view.
That one is a special kind of mess, will be fixed separately.

GitOrigin-RevId: fbac1892082696a6c5b36df5d91a379c35dd63e8