IDEA-320506 Fix details in Project View for Packages View
This turned out to be more complicated than it seemed.
The problem is that it uses PackageElementNode which
is a part of Java PSI implementation that isn't even available
in lang-impl. And it has no useful inherited methods either.

So we employ a possibly ugly trick here: instead of checking
that the parent node is a directory, we check that it can't be
a file instead. After all, all this code is about avoiding inplace
comments for members, and members are usually a part of
something backed by a virtual file which is a real file and not
a directory.

GitOrigin-RevId: 1e5fa26d2cebe343d7b279713285e45467b2b69f