[runtime module repository] remove unnecessary ../ prefix from resource paths in module descriptors (IJPL-109)
Module descriptors are put to a separate module-descriptors.jar now, and we don't plan to place their resources to that JAR, so the paths will always start with '../' prefix, and it'll just complicate code and requires additional work at runtime. So let's suppose that paths in module-descriptors.jar are relative to the directory where the JAR file is located.

GitOrigin-RevId: 43572247998686fd81d66bcdfba8dd700605281b