[kt-*-only] Use Kotlin JPS plugin 1.9.0-Beta-27 for local development
On build server in kt-*- branches jps plugin and compiler
are both equal to the last Kotlin version.
This commit is required for tests (MPP Import tests in particular).
Without it, execution on build server and locally produces a
different testdata, because at some point we are using the
default LV which is determinate by the bundled JPS version

GitOrigin-RevId: 2d211fd71f3c1a1766b93cbcc8f6c54a1f1947b8