IDEA-319836 Don't update tool window weights if editor area isn't there
Sometimes, apparently when the IDE is being initialized, some tool windows
appear before the editor area (EditorSplitters technically) is created and visible.
When that happens, ThreeComponentSplitter uses that area to extend the tool
window beyond its weight. In the worst case the Project View takes all available
space. Once its weight is updated based on its new size, it just stays like that
forever.

To fix that, we check that the middle component is present and visible. If it's not,
we don't update weights and write a log message. While this doesn't fix the original
issue, it prevents the new weight from sticking and the Project View taking all space
until manually resized.

While we're at it, add some useful logging when weights are updated or not updated.

GitOrigin-RevId: 7e03dc699b569940db1706c69f474c1f5df4caf2