Disallow Java in-place rename in non-Java files
The Java in-place renamer currently allows in-place renaming for members
in any file where a reference appears. But this doesn't always work when
the reference to the member is in a non-Java file.

This change prevents Java's in-place rename from running when the
reference being renamed is not in a Java file. The target language still
has the chance to provide their own renamer if desired.

This change addresses https://youtrack.jetbrains.com/issue/IDEA-321110