[runtime module repository] cache classpath for intellij.platform.bootstrap module in MANIFEST.MF (IJPL-128)
Currently com.intellij.platform.runtime.loader.Loader parses all module descriptor xml files to compute classpath for ModularMain class, this may take noticeable time. Now classpath for the bootstrap module is cached in an attribute of MANIFEST.MF file, so it can be loaded faster. This is done for production mode only, because in that mode paths to classes roots don't use macros, so there is no need to substitute them.

GitOrigin-RevId: 0af716f718976b67c3a6bcfc94358a0b488d8649