Replace CancellablePromise and a pooled thread executor with Coroutine in CodeStyleCachedValueProvider
In order to improve CPU utilization, this commit replaces the use of a pooled thread executor with coroutine in the CodeStyleCachedValueProvider.kt file. The most notable change is the start() function's implementation.

The start() function now launches a coroutine and properly follows the pattern of calling readAction and then dispatching the result to the EDT using dispatchers and withContext. The cancel() function has also been updated accordingly to cancel launched jobs.

Also, CodeStyleCachedValueProviderService was added as a project-level service for providing coroutineScope to the provider. This service will be used to manage and properly cancel coroutines related to CodeStyleCachedValueProvider.

These changes ensure that coroutines are used instead of promises for managing code style computation tasks, providing a more standardized and readable approach.

GitOrigin-RevId: 120e9968180ff19f4415933775cf3bea6a1683e7