EA-789902 Make CustomJBDateTimeFormatter thread-safe
When the user has overridden the date format in the settings,
then CustomJBDateTimeFormatter is used instead of JBDateTimeFormatter,
but the former, unlike the latter, is not synchronized.

Since SimpleDateFormat is not thread-safe, synchronization is necessary,
otherwise it may lead to obscure bugs and exceptions as the format is mutable
internally. Looks like it's the most likely cause of this EA issue because
otherwise it would mean that there's a weird math bug in the date format,
which is far less likely.

Another way to solve this would be to use thread locals, but since we already have
this SyncDateFormat class that is used in the default format which is used for
most users with no noticeable issues, it seems that reusing it here is a good idea.

This change caused protected signatures to change, but since they're not
used elsewhere, internally or externally, we might as well make them private,
just like Qodana tells us to. And make the formatter itself private too.

GitOrigin-RevId: e5fe52fb6ccfae2239009ab1225d078e2d9306d5