IDEA-321129 Hide tool windows with no saved state when restoring layout
The old behavior: if the saved layout contains no information about
a tool window (because it wasn't present when the layout was saved,
a very common case), then its current state is simply preserved.

The new behavior: if the saved layout contains no information about
a tool window, copy its current state, but set isVisible to false, so
that it's at least hidden when the Restore Tool Window Layout
action is performed.

This is a more consistent behavior for users, as they expect to see
the same set of visible tool windows as it was when the layout
was saved.

GitOrigin-RevId: c49a95e9b69da2deb396edabf13f97d0ae40e8b3