(IDEA-321185) ML in SE: add custom reordering for items from semantic contributors
Currently, the ranking model was not trained on data with cosine similarity as a feature. That means the model cannot account for it when calculating ML priority and may lead to semantic candidates appearing at the end of the list with results. To solve this, I manually order semantic items by similarity score and place them between classic items starting from the third position.

GitOrigin-RevId: 49e0ddd3d8d7e09153da4f615673bef6d46f5082