Make WebpImageReaderSpi, WebpImageWriterSpi, and WebpWriter final
Refactor WebpImageReaderSpi and WebpImageWriterSpi to be final classes, as they do not need to be extended further. Also, change the constructor of the nested WebpWriter class in WebpImageWriterSpi to be private, keeping it more encapsulated. Remove unnecessary import in WebpMetadata and change the mergeTree method to not throw IIOInvalidTreeException since it is never used. These changes increase code-readability and maintainability.

GitOrigin-RevId: 68c6af827da87d9dac475595e80a1b31219799af