remove deprecated addProblemListener and refactor code
This commit removes the deprecated `addProblemListener` method in `WolfTheProblemSolver.java` as it advised using message bus `ProblemListener#TOPIC` instead. The `MockWolfTheProblemSolver.java` is also modified to remove the overridden deprecated method. In `WolfTheProblemSolverImpl.java`, the usage of `addProblemListener` is replaced with a direct connection to the message bus.

Additionally, this commit includes some minor refactoring and code style changes, such as updating method return types to use the `@NotNull` annotation, making the `ProblemFileInfo` class final, and handling cases where the document may be null.

GitOrigin-RevId: f88f88aae302ef20dbf31173ea4b3fee909b8ee2