IDEA-321235 Update saved IdeFrame bounds if moved when maximized
The problem: the frame's bounds were only saved if the frame is not
in the maximized state when it's moved from one screen to another.

To solve it, store not just the last "normal" (non-maximized) frame bounds,
but also the bounds of the screen they correspond to.

Then, when the frame is moved in the maximized state, we move/scale
the last normal bounds if the current screen is different, and save them.

This way when the IDE is started we always have the normal bounds
corresponding to the screen the IDE frame was on when it was closed.
If it was maximized, we still use those bounds to choose the right screen.

GitOrigin-RevId: 38f607bff9ec1ea24d69879cc8ee2daff67e5930