KTIJ-25670 [devkit] Migrate CompanionObjectInExtensionInspection to Analysis API
This way, the inspection does not throw exceptions
when executed in K2 IDE.

Important: Since the quickfix relies on a "Move refactoring"
implementation, it will most probably not work in K2 IDE for
now, and this issue should be addressed separately.

Also, this change fixes a false positive caused by
type aliases (see IDEA-321279), and this commit adds
additional tests for that.

^KTIJ-25670 Fixed
^IDEA-321279 Fixed

GitOrigin-RevId: be7b2ac4a515bd381ea20084ef63fb0cee12f0f8