CPP-33475: Create the Qt plugin
Split a part of QML plugin with core functionality for paths detection and
with icons specific to Qt file types.
For now the functionality is used in CLion and Python plugin.

This change allows to open `.ui` files with detected `designer` tool which didn't
work in CLion before.
The icons are also updated to the more up-to-date format which honors new UI.

IJ-CR-107960

GitOrigin-RevId: dffc1706197ebc568db43ff85f324c1520e919f6