[Workspace Model] Update an implementation of the entity reference to use storage instrumentation
In the previous implementation of the entity reference we had a cast to the entity storage implementation and called some internal functions of it.
This is not scalable for the case of tracked entity storage. If our storage is wrapped with a traced proxy, the cast failed. Now the reference resolve is located in the instrumentation level and can be used with proxy.

GitOrigin-RevId: 3f7b888933eb0e3869daf2367d58dc7e18af2a00