[vcs-log] do not create default columns when graph table is created
This can cause NPE for columns from the unloaded plugins, which ids are kept in the modelIndices map but are not supposed to have actual columns created for them.

Since default columns are not created, when invalid column order is loaded from the properties, current columns in the table may not be available to use as a fallback. In this case, Root and Commit columns are added to the saved order manually.

EA-322577

GitOrigin-RevId: 1d6f4ef6ace1ecafcede32911c4c8b8b6e73e727