IDEA-321388 Log more IDE frame events related to updating bounds
com.intellij.openapi.wm.impl.FrameInfoHelperKt.updateFrameInfo
is very complicated, so log as much stuff as possible there to
have a chance of figuring out what went wrong and why.

Also log reasons for NOT updating frame bounds / screen:
either because the current bounds are unknown or because
the screen is the same or has invalid bounds.

GitOrigin-RevId: 67949e26cf906b6b89474e7752222ccdc88c61ce