[Feedback] Introduce In-IDE Feedback Surveys API
The goal of this API is to allow plugins to easily create and perform feedback polls inside and outside the IDE.

For now, this API will only work for JetBrains plugins. In the future we may expand availability for all plugins (or not).

One important API goal I would like to achieve is to strictly limit the types of surveys (com.intellij.feedback.common.openapi.FeedbackSurveyType) and for each type define a specific behavior that can be customized in a limited way using a configuration object (com.intellij.feedback.common.openapi.FeedbackSurveyConfig).

IJ-CR-110796

GitOrigin-RevId: 3f3b0c25f5bd85d0c073db01120f823b5ae1e587