[debugger][kotlin] IDEA-319901 Log some context of an evaluation error to simplify reproduction
Log project, file name where the program is suspended at, and a suspension line. Attach the evaluated expression and file contents (not included by default for external users) as well.

As we expect fewer bug reports now, when we automatically fall back to the old evaluation, such context should help us address the cause of the issues.

GitOrigin-RevId: fafd4dc3875f373438100b4e0ff8c358215b4604