[platform] fix ModuleUtilCore::findModuleForPsiElement for files from custom external entities (IDEA-321398)
If there is a custom entity with WorkspaceFileIndexContributor which excludes a root from a module content and registers it as an external root, we except that findModuleForPsiElement returns null for files under that root. However, ProjectFileIndex::getOrderEntriesForFile may return ModuleSourceOrderEntry instance for such files since RootIndex doesn't use information from WorkspaceFileIndex, so we need to explicitly take only LibraryOrSdkOrderEntry instances into account.

GitOrigin-RevId: 2ac7012f247ae9ddf4184412e4c220c389e8c293