IDEA-319986 DefaultTreeLayoutCache: new, efficient implementation
This should fix all VariableHeightLayoutCache performance issues
once and for all, at least when it's polished and bug-free, which
it isn't yet. Once it's done, however, it's intended to replace both
VariableHeightLayoutCache and FixedHeightLayoutCache,
as the latter doesn't play well with async trees. This one is
based on VariableHeightLayoutCache, so it works fine with async
trees.

The new implementation addresses the following issues:

1. Excessive bounds calculations involving a lot of heavyweight
font metrics related stuff, potentially for hundreds of thousands
of nodes. Addressed by only computing bounds on the fly.
In case of variable height, heights of invisible but not yet
computed nodes is assumed to be equal to default,
which doesn't cause any serious issues.

2. Inefficient algorithms are optimized. For example,
the row index of a node was computed using linear search.
If that happened in a loop, that led to O(N^2) complexity.
Addressed by caching row indices and updating them
when the visible structure changes, which has linear complexity.

3. Variable height is dealt with by using a Fenwick tree to
store the sum of differences between the default and
the actual height of every node. Whenever the height of a node
is computed or recomputed, the difference can be updated
in logarithmic time instead of updating the whole tree in linear time.
To find a node by its Y coordinate, binary search is used,
which looks up the Y coordinate at every step using the Fenwick tree.
This leads to quadratic logarithmic complexity, which is still
very fast - hundreds of operations for trees with millions of nodes.

GitOrigin-RevId: 13b5682076b210336c17b15e71084002d1157d1d