[workspace model] IDEA-322007 Library doesn't update if pulling a commit with changed version and more then one project opened
The problem is easily reproduced if you have two open projects.
The problematic case:
 After the `git pull` command for the commit with the changed library, it was not updated until the project was restarted.
The problem was in reading outdated data from `StateStorageBase#getStorageData`. Usually `JpsProjectModelSynchronizer` was notified by `StoreReloadManagerImpl#applyProjectChanges` after some VFS events that we need to reread the content of the files. But due to the fact that `StoreReloadManagerImpl` was rewritten from application-level service to project-level it still continues to handle the logic for all opened projects. And we faced the case then StoreReloadManager for project B based on the event for project B notified JpsProjectModelSynchronizer for project A that it needs to reload its state but data for StoreReloadManager A was not actually updated yet (and we read old data).  The fix is simplest,  StoreReloadManager should notify only `JpsProjectModelSynchronizer` related to its project, not others.

GitOrigin-RevId: 8e707b3925ca682ed82c04d388b23a62f1e36cac