IDEA-315623 make modal progress resistant to failures in Swing hierarchy
Before the change the error in Swing internals caused the task cancellation leaving the IDE in an unusable state.
After the change, the task is run without cancellation and the modality is still handled correctly.
The used might not see the modal progress, and the error will be logged, but at least the IDE will not brick itself.

Example:
```
java.lang.NullPointerException: Cannot invoke "javax.swing.JRootPane.getGlassPane()" because the return value of "javax.swing.JFrame.getRootPane()" is null
	at java.desktop/javax.swing.JFrame.getGlassPane(JFrame.java:718)
	at com.intellij.openapi.ui.impl.GlassPaneDialogWrapperPeer.createDialog(GlassPaneDialogWrapperPeer.java:104)
	at com.intellij.openapi.ui.impl.GlassPaneDialogWrapperPeer.<init>(GlassPaneDialogWrapperPeer.java:85)
	at com.intellij.openapi.progress.util.ProgressDialogWrapperKt.peerFactory$lambda$1(ProgressDialogWrapper.kt:102)
	at com.intellij.openapi.ui.DialogWrapper.<init>(DialogWrapper.java:318)
	at com.intellij.openapi.progress.util.ProgressDialogWrapper.<init>(ProgressDialogWrapper.kt:25)
	at com.intellij.openapi.progress.util.ProgressDialogWrapperKt.createDialogWrapper(ProgressDialogWrapper.kt:87)
	at com.intellij.openapi.progress.impl.PlatformTaskSupportKt$showModalIndicator$1$1.invokeSuspend(PlatformTaskSupport.kt:265)
	at kotlin.coroutines.jvm.internal.BaseContinuationImpl.resumeWith(ContinuationImpl.kt:33)
	...
```

GitOrigin-RevId: c03b71ee62870b805c933c8a591e867d4c4b6174