IDEA-299506 Fix potential infinite loop in DefaultTreeUI
It should not be possible for the bounds of a tree path having a valid row
to be null. However, experience shows that it happens sometimes, and
it's definitely better to ignore the invalid row's width than to enter
an infinite loop just because of a stupid bug in the tree layout cache.

GitOrigin-RevId: 9833e2987364d6ad97a7ddc041e347d6dbcdcd23