[workspace model] IDEA-322007 Get rid of one more possible case when the project model can't be updated
If for some reason will be a delay between VFS events processing in `StorageVirtualFileTracker`, we will face the situation when `StoreReloadManagerImpl#doReload` will be called, and it will reload only part of the events.  After that, `JpsProjectModelSynchronizer#reloadProjectEntities` will be called that know all VFS events. In such cases, we will read only partially updated data.
To avoid this possible problem in the future, let's process all events in `StoreReloadManagerImpl` at once and only after that calls `JpsProjectModelSynchronizer`.

GitOrigin-RevId: f370bc91b2daed1f597a3834a25cc5b3d04eb32e