IDEA-274904 Fix broken stripe button dragging
Turns out we can't just disable ToolWindowDragHelper completely
because it's also responsible for dragging stripe buttons.

So instead of just returning false right away, move the new
code to getToolWindowAtPoint() and make it return null if it's
the header that's being dragged and header dragging is disabled.

This function is called in one other place, where it stores a weak
reference to the window being dragged. It's totally fine to return
null there as well because the window is not going to be dragged
anyway if the setting is off, so what's the point of storing that
reference?

GitOrigin-RevId: 3b897c4d67b94fb875bfd45e52804740c5115582