Revert "Don't load the bundled scheme from the same resource path twice: IDEA-321940"
The commit led to a critical problem IDEA-322405: when the SchemeManager is reloaded, it removes all schemes, but does not clean the newly introduced collection, therefore a scheme is returned which is not contained in the SchemesCollection. A more correct fix of IDEA-321940 will be made later.

This reverts commit 301998ec385a7f92582c2c7db224e15f5df1ec1b.

(cherry picked from commit 02d2437b9e18b5ed85ef9fababb3f9c21ad4278f)

IJ-CR-108605

GitOrigin-RevId: 1d5af52ba57a8ce1e1387e0e6a976045cd8d53b7