IDEA-308676 Increase repaintPath area in DefaultTreeUI
This is an ugly fix for painting artifacts caused by fractional scaling on Windows.
A proper fix would be to either perform a lot of complicated computations
to figure out the exact user-space coordinates or, more globally,
get rid of JBR-side fractional scaling altogether, which is nowhere near
any planning horizon at the moment.

GitOrigin-RevId: 39da3b910a7c1ab41562707353a7080122d45a96