[Kotlin] MPP: detect test compilations by associated compilations instead of name
The previous name-based heuristic won't support upcoming versions of AGP,
since there will be no dedicated platform for Android in the future.

For corner cases an import-specific Gradle property is provided.
It rolls back import logic to the old style. It can be used for
edge cases, such as creating a production compilation that has
to be associated with the main compilation.

KTIJ-25395

GitOrigin-RevId: 97001040a146f62ac8c84798f5e0fb86eadde1ec