RUBY-31394 always execute WSL commands in a real/user shell.
Previously, when shell path was not specified, it defaulted to `/bin/sh`.
This caused the following problems:
- we did not respect user's choice of login shell.
- we lost part of the user's environment, if user didn't use /bin/sh as their login shell.
- different commands were executed with different shells (i.e. different environments), which lead to inconsistent execution results.

Now we do not allow to specify the shell path at all and always use user's login shell.
`/bin/sh` is still used as a backup for cases when user's login can't be determined for some reason.

Reviewed-by: IJ-CR-108024

GitOrigin-RevId: 357187dd7ab4e08e0d87ff3d36d6bf1a33f83245