[debugger][kotlin] IDEA-319901 Report exception only from internal users
The IDE collects a handful of sensitive aspects of the execution to ensure it's possible to try and reproduce the exception. So, some attachments, like the code, are by default excluded from the report. If an external user wants to share such details, they should manually set the checkbox in the IDE Internal error form.

However, even in that case, we may not have a chance to reproduce the exception, as we may not have access to the external code, let alone means to recreate the same execution. So all that comprehensive and sensitive information may give us nothing to work with.

Many internal users work with Kotlin, and we receive many exceptions from them, so this group should be enough to sort out possible problems.

(cherry picked from commit da517ead64d0b3427d4addb76d2ff1d71c9ed5f8)

GitOrigin-RevId: 7b84e8f7b69a4e4a070836d2ee7ce8f6928de94e