[kotlin] UsePropertyAccessSyntaxInspection: check all declarations for non-trivial body
There are cases when the descriptor resolves to multiple declarations in the IDE,
one of which has no body (ex. interface), and another has a non-trivial accessor body.

In that case we conservatively don't report a warning.

I failed to come up with a test for this, but reproduced it in intellij repo
with intellij->kotlinc->intellij dependency chain.

^KTIJ-25858

GitOrigin-RevId: 71cce8fcd6f252f44ec34d5c0757fe3c911a9d5f