[rdct] fix config synchronization when the client started for the first time (RDCT-6)
Before, if the config folder didn't exist when the process starts, synchronization wasn't enabled, so if a new client process was started later, changes in configuration made in it won't be visible in the first client. Now watching for changes in config files is enabled in such cases. Also, ThinClientPathCustomizer supports the custom config import marker file, and it's used in tests to properly emulate the first run of the client.

GitOrigin-RevId: a2d11476816bd3c95dd18dae752827885d6bc78b