[debugger][kotlin] IDEA-319901 New registry key for the JVM IR evaluator
Now that we have the fallback from the new JVM IR evaluator to the old one and keep track of errors occurred with the new evaluator, it's beneficial for both sides to have the new one enabled by default.

The old key was exposed and turned off by many users, including internal ones. Reaching every single one of such users is nearly impossible, plus it requires additional actions from them.

Another approach would be to just make the old key do nothing, so everyone uses the new evaluator regardless of what's set there. Whereas it solves the problem, there are times when it's desirable to use the old evaluator — for example, to test it (well, fallback does exist, so it means the old evaluator still has work to do). And, in case of an extra emergency, users still have means to disable the new evaluator.

GitOrigin-RevId: 8c11afd02370912fdcfe284d725d9392290e65dd