KTIJ-25601 Script configurations not loaded at kotlin-lang project
The issue happened due to data race. There are two concurrent processes
taking place at project opening:
(A) Script configurations loading [1].
(B) Editor notifications collection [2].

The second activity is based on the assumption that script configurations are
loaded and can be safely used. Missing configuration is an exceptional situation.
Due to the race this statement was sometimes false positive.

`GradleBuildRootsLocator.findScriptBuildRoot()` from time to time could not find
`GradleScriptInfo` associated with the script file (loading was not yet over).
Because of that `GradleBuildRootsLocator.ScriptUnderRoot.isImported` was `false`.
And hence the wrong value of `GradleBuildRootsLocator.ScriptUnderRoot.getNotificationKind`:
`NotificationKind.notEvaluatedInLastImport` instead of `NotificationKind.dontCare`.

To fix the issue this commit introduces new API call [3] allowing to get
configuration cache considering the end of its load.

----------------------------------------------------------------------
[1]: init-block of ScriptClassRootsUpdater
[2]: GradleScriptNotificationProvider.collectNotificationData
[3]: ScriptClassRootsUpdater.getNotEmptyClassPathRoots

^KTIJ-25601 fixed

GitOrigin-RevId: d1d9f7934f2830663b6ac4014dcae55cf5e4e237