PY-60959 During skeleton generation skip module attributes with illegal names
Otherwise, we end up generating syntactically invalid skeletons or, if
a large module gets transformed into a package where each class stub is
declared in its own file with the same name, some of the class names might
be illegal even for the target file system.

In this particular issue, the "shyft" package declared a class named
"ComputeNodeInfoVectorA strongly typed list of ComputeNodeInfo as returned
from the dstm server\012", which couldn't be copied to a Windows host from
a Linux remote.

IJ-MR-109009

GitOrigin-RevId: 17bd63bc1c3d1925880161f5b801724492cfb038