IDEA-320139 Create ProjectViewPerformanceCollector to measure dir expansion
This is a tricky metric for two reasons.

1. There may be many causes for a directory expansion. For example,
when a project is loaded, the tree expansion state is restored,
which causes a lot of expansions. They shouldn't be measured
because there is a lot of them and because too much stuff is going
on at that time. Another example is the Expand All action, which
can produce thousands of expansions.

2. Due to AsyncTreeModel, expansion happens in background,
so we must subscribe to model changes and track them very
carefully. That is, unless the children were previously loaded.

The first problem is solved by creating the Tree.startMeasuringExpandDuration
internal method, which is called before expansions that should be measured.
At this time it's DefaultTreeUI-originated expansions (toggleExpandState)
and the selectChild action (right arrow key).

The second problem is solved by creating a dedicated class,
ProjectViewDirectoryExpandDurationMeasurer, which subscribes to the model
if the 'loading...' node is shown and then carefully tracks events.
If an event removes the dir being tracked, the measurement is cancelled.
If it's an expansion event, the measurement is finished.

GitOrigin-RevId: bfdc85b5c4dc9dca6ef466ee01b695bced012a7c