PY-60959 Restart code analysis after an SDK update even if introspection has failed
The reasoning is that if an interpreter can't be successfully updated, it's better
to fail fast, clearly indicating the current state of resolve, rather that confuse
a user by invisible errors. It might go the other way around, e.g. the update has
been mostly completed before an error occurred, but we still show unresolved references
in the editor, even though all the remote sources and skeletons are already available
for code insight.

In this particular case, a runtime SftpChannelException, caused by an attempt to copy
a file with an illegal name, made highlighting go stale and still report unresolved
imports until the IDE was restarted.

IJ-MR-109009

GitOrigin-RevId: ee218dd575205de054e7d093ac6635d168449120