cancelAndJoinBlocking from shutdown.kt: Write coroutines dump inside dedicated thread
Sometimes it happens that Dispatchers.Default is fully occupied, especially in tests, especially when something is stuck. In that case, the debugging coroutine, which is supposed to work when everything else doesn't, shouldn't use the limited default thread pool.

The most reliable solution is to launch the coroutine in a certainly unlimited thread pool.

Merge-request: IJ-MR-108735
Merged-by: Vladimir Lagunov <vladimir.lagunov@jetbrains.com>

GitOrigin-RevId: f14bd8f5073fa6ef754f606fdbf79ab803a85ba0